+++ app/src/main/AndroidManifest.xml
@@ -16,19 +16,19 @@
         android:theme="@style/Theme.PlayListMaker"
         tools:targetApi="31">
         <activity
-            android:name="com.practicum.playlistmaker.presentation.MediaActivity"
+            android:name="com.practicum.playlistmaker.media.presentation.MediaActivity"
             android:exported="false" />
         <activity
-            android:name="com.practicum.playlistmaker.presentation.SearchActivity"
+            android:name="com.practicum.playlistmaker.search.presentation.SearchActivity"
             android:exported="false" />
         <activity
-            android:name="com.practicum.playlistmaker.presentation.SettingsActivity"
+            android:name="com.practicum.playlistmaker.settings.presentation.SettingsActivity"
             android:exported="false" />
         <activity
-            android:name="com.practicum.playlistmaker.presentation.PlayerActivity"
+            android:name="com.practicum.playlistmaker.player.presentation.PlayerActivity"
             android:exported="false" />
         <activity
-            android:name="com.practicum.playlistmaker.presentation.MainActivity"
+            android:name="com.practicum.playlistmaker.main.presentation.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
