--- app/src/main/java/com/practicum/playlistmaker/data/ItunesResponse.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmaker.data
-
-import com.practicum.playlistmaker.domain.Track
-
-data class ItunesResponse(val results: List<Track>)
