+++ app/src/main/java/com/practicum/playlistmaker/main/presentation/TrackAdapter.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.presentation
+package com.practicum.playlistmaker.main.presentation
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,7 +7,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.practicum.playlistmaker.domain.Track
+import com.practicum.playlistmaker.search.domain.Track
 import com.practicum.playlistmaker2.R
 import kotlin.collections.ArrayList
