+++ app/src/main/java/com/practicum/playlistmaker/media/presentation/MediaActivity.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.presentation
+package com.practicum.playlistmaker.media.presentation
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
