+++ app/src/main/java/com/practicum/playlistmaker/search/data/ItunesResponse.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.search.data
+
+import com.practicum.playlistmaker.search.domain.Track
+
+data class ItunesResponse(val results: List<Track>)
