+++ app/src/main/java/com/practicum/playlistmaker/search/domain/NetworkService.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.domain
+package com.practicum.playlistmaker.search.domain
 
 interface NetworkService {
