+++ app/src/main/java/com/practicum/playlistmaker/search/domain/Track.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.domain
+package com.practicum.playlistmaker.search.domain
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
