+++ app/src/main/java/com/practicum/playlistmaker/search/domain/TrackSearchCallback.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.domain
+package com.practicum.playlistmaker.search.domain
 
 interface TrackSearchCallback {
